+++ firebase.ts
@@ -13,5 +13,6 @@ const firebaseConfig = {
 
 const app = firebase.initializeApp(firebaseConfig)
 const db = app.firestore()
+const auth = firebase.auth()
 
-export { db }
+export { db, auth }
