+++ src/pages/landing/index.tsx
@@ -1,0 +1,7 @@
+export default function Landing() {
+  return (
+    <div>
+      <h1>WHATS GOOD IN DA NEIGHBORHOOD</h1>
+    </div>
+  )
+}
