+++ src/pages/login/index.tsx
@@ -1,15 +1,18 @@
 import { useState } from 'react'
-// import { db, auth } from '../../../firebase'
-import { createUserWithEmailAndPassword, getAuth } from 'firebase/auth'
+// import { db } from '../../../firebase'
+import { useRouter } from 'next/router'
+import { auth } from '../../../firebase'
 
-const SignUpPage = () => {
+const LoginPage = () => {
   const [email, setEmail] = useState('')
   const [password, setPassword] = useState('')
-  const auth = getAuth()
+  const router = useRouter()
 
-  const handleSignUp = async () => {
+  const handleLogin = async () => {
     try {
-      await createUserWithEmailAndPassword(auth, email, password)
+      console.log(process.env.API_KEY)
+      await auth.signInWithEmailAndPassword(email, password)
+      router.push('/landing')
     } catch (error) {
       console.log(error)
     }
@@ -17,7 +20,7 @@ const SignUpPage = () => {
 
   return (
     <div>
-      <h1>Sign Up Page</h1>
+      <h1>Login Page</h1>
       <form>
         <label>
           Email:
@@ -35,12 +38,12 @@ const SignUpPage = () => {
             onChange={(e) => setPassword(e.target.value)}
           />
         </label>
-        <button type="button" onClick={handleSignUp}>
-          Sign Up
+        <button type="button" onClick={handleLogin}>
+          Login
         </button>
       </form>
     </div>
   )
 }
 
-export default SignUpPage
+export default LoginPage
